Samples V0.0.1
Samples have a unique name and a description
Sample have parameters
Only one sample can be active for an experiment at a time
A new run gets associated with the sample that is "current" at the start
of the run.
We can see the sample in the user run tables.