Move methods that change data to POST's.
There's a few methods where I incorrectly used GET's; these should be
POST's and in the rare case, DELETE. I have removed the GET support for
these in order to support backwards compability for a bit but will
eventually remove GET support. All parameters are still in the query
string; so we'll have do some extra work in the ajax request.

This is in preparation for locking down an experiment.
I intend to use the HTTP method to determine if the endpoint changes
data.