Mechanisms to edit the system run tables.
We can now clone run tables.
We clone run tables into an experiment, edit/save and then we "publish"
by marking the run table as a system run table.

All of this uses a site_admin bool that get's set in the JS.

Also, add a method to accommodate the calibration service.
Return run info similar to the old logbook.