Post to instrument elogs.
Add a "elog" instrument parameter with the name of the instrument
specific instrument elog.

Add an array of experiment spanning elogs to an "info" document in the
site database for additional elogs. For example,
use site
db["info"].insertOne(
  {"experiment_spanning_elogs": [
   "Photon Systems",
   "Sample Delivery"
  ]})

We do the best to update the instrument elog on elog followup's, updates
and deletes. However, not all operations are copied over.