Create experiment with initial sample does not rebuild the cache.
In Cryo, samples are validated; an exception is thrown if the sample 
does not meet the sample criteria. This exception causes the server to 
skip publishing the Kafka message.

Workaround: If the server says "The experiment" already exists but the 
experiment is not visible; then this is a potential culprit. 
Refresh/rebuild the experiment caches and the experiment should show up 
again.