Clean up backup..
Giving up on incremental backup for now.
Can't figure out a foolproof way to determine if a database has changed.
We may need to do something like process the application logs and so on.
While that seems hacky, it's probably more reliable than any other scheme.
And it's probably applicable to calib/config/captar as well.
So for now, we follow Knuth's advice and do not prematurely optimize. That comes later.