Publish messages locally.
This is a tricky update. In addition to publishing messages on Kafka, we
also add to a local queue for the explgbk_cache. We use this local queue
to rebuild the experiment info in explgbk_cache. This may have
unintended consequences; so we need to test this out properly in a
clustered environment.