Attempt 2 at solving Cryo's experiment refresh issue.
Trigger update for the "experiments" topic from within Javascript.
This is done on creating/editing/cloning an experiment.
Because we get two messages, if there is a bug, we could potentially see two entries for the same experiment (pointing to the same object tho).
As a side effect, we refresh the tabs for an instrument when switching instrument tabs in "Select Experiment".